Welp, this time it was software. Fixed acq home, fixed demo mode logic
A confusing naming scheme (drives.toTele(true) is demo yet the TELEBASE
state is demo/general while TELEOP and drives.toTele() is normal, little
un-intuitive) and a missing else statement in init led to a sequence of
"fixes" that broke otherwise good logic below, set the wrong state, and
eventually created essentially a double flipped boolean. My bad. Looking
through commit history for translation of fixes across all branches has
revealed the unfortunate self code-reading error. The acq home was
actually just a simple accidental commenting of the a button in
buttonState setting, and a climb-disable button is now back in as well.